Overhaul release blessing process.
Blessing is now the time that bootstrap tarballs are created.

There's a banner that goes into the global manifest (stored in
scripts/admin/banner.txt).

The global manifest.json is now entirely generated: there's no record of the
current version stored in git.

Have not yet updated launch-meteor or install-template.sh or the fake release
for this brave new bootstrap world, or https://install-bootstrap.meteor.com/.