make TEST_WAREHOUSE_DIR a non-test thing called METEOR_WAREHOUSE_DIR
(the non-springboarding piece of that is now controlled by
METEOR_TEST_NO_SPRINGBOARD)

This means that run-all-tests.sh can use launch-meteor to bootstrap a warehouse
instead of starting with an empty warehouse directory (which is no longer
allowed).

also, fix npm module usage out of the bootstrap tarball: we were too liberal in
the "never update npm dependencies in the warehouse" rule since the bootstrap
tarball would contain packages with no node_modules (or if we had a failure
between downloading a package and running installNpmDependencies).  instead,
"never update npm dependencies in the warehouse if .npm/node_modules already
exists".